# docs/swagger.yaml
openapi: 3.0.3
info:
  title: UploadHaven API
  description: |
    UploadHaven is a modern, secure file sharing platform built with Next.js 15, TypeScript, and MongoDB. 
    This API provides comprehensive file management, real-time notifications, security scanning, and analytics capabilities.
    
    ## Features
    - 🚀 Secure file upload and sharing
    - 🔔 Real-time notifications via Server-Sent Events
    - 🔒 Malware scanning and security monitoring
    - 👥 User and admin management
    - 📊 Analytics and statistics
    - 🎯 Rate limiting and abuse prevention
    
    ## Authentication
    The API uses session-based authentication with cookies. Most endpoints require authentication.
    
  version: 2.0.0
  contact:
    name: UploadHaven Support
    url: https://github.com/yourusername/uploadhaven
    email: support@uploadhaven.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://uploadhaven.vercel.app
    description: Production server
security:
  - sessionAuth: []
...
paths:
...
  /api/upload:
    post:
      tags: [Files]
      summary: Upload a file
      description: Upload a file with optional password protection and expiration
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload (max 100MB)
                expiration:
                  type: string
                  enum: ["1h", "24h", "7d", "30d", "never"]
                  default: "24h"
                  description: File expiration time
                userId:
                  type: string
                  description: User ID (must match authenticated user)
                password:
                  type: string
                  description: Password protection for the file
                autoGenerateKey:
                  type: boolean
                  default: false
                  description: Auto-generate a secure password
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Invalid file or parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
...
  /api/security/scan/file:
    post:
      tags: [Security]
      summary: Scan file for malware
      description: Scan a specific file for malware and security threats
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileName]
              properties:
                fileName:
                  type: string
                  description: Filename to scan
                  example: "document.pdf"
      responses:
        '200':
          description: Scan completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResult'
        '404':
          description: File not found
        '500':
          description: Scan failed
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication using HTTP-only cookies
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: string
    RateLimitError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            rateLimit:
              type: object
              properties:
                limit:
                  type: integer
                remaining:
                  type: integer
                reset:
                  type: string
                  format: date-time
...
    UploadResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        url:
          type: string
          example: "http://localhost:3000/uploads/public/abc123.pdf"
        shortUrl:
          type: string
          example: "http://localhost:3000/s/abc123"
        filename:
          type: string
          example: "abc123.pdf"
        originalName:
          type: string
          example: "document.pdf"
        size:
          type: integer
          example: 1048576
        type:
          type: string
          example: "application/pdf"
        expiresAt:
          type: string
          format: date-time
        generatedKey:
          type: string
          description: Auto-generated password (if requested)
        metadata:
          type: object
          properties:
            id:
              type: string
            uploadDate:
              type: string
              format: date-time
...
    ScanResult:
      type: object
      properties:
        fileName:
          type: string
        scanResult:
          type: object
          properties:
            isMalicious:
              type: boolean
            isSuspicious:
              type: boolean
            threatName:
              type: string
            source:
              type: string
              enum: [local, virustotal]
            scannedAt:
              type: string
              format: date-time
            details:
              type: string
        quotaStatus:
          type: object
          properties:
            remaining:
              type: integer
            used:
              type: integer
            limit:
              type: integer
            resetTime:
              type: string
              format: date-time
        scannedAt:
          type: string
          format: date-time
...